Move the label parser into the library so every front end shares one implementation
